refactor(Optim): refine optimizer codes
