+++ oneflow/python/nn/optimizer/adam.py
@@ -21,8 +21,7 @@ import oneflow as flow
 
 from oneflow.python.oneflow_export import oneflow_export
 from oneflow.python.nn.parameter import Parameter
-from oneflow.python.nn.optimizer.optimizer import ParamGroup
-from oneflow.python.nn.optimizer.optimizer import Optimizer
+from oneflow.python.nn.optimizer.optimizer import Optimizer, ParamGroup
 
 
 @oneflow_export("optim.Adam")
@@ -106,14 +105,8 @@ class Adam(Optimizer):
             for param in param_group.parameters:
                 assert param.is_leaf, "parameters must be leaf tensor"
                 self._state[param] = dict()
-                self._state[param]["exp_avg"] = flow.tmp.zeros(
-                    # TODO: zeros module support flow.Size parameter
-                    tuple(param.shape)
-                )
-                self._state[param]["exp_avg_sq"] = flow.tmp.zeros(
-                    # TODO: zeros module support flow.Size parameter
-                    tuple(param.shape)
-                )
+                self._state[param]["exp_avg"] = flow.tmp.zeros_like(param)
+                self._state[param]["exp_avg_sq"] = flow.tmp.zeros_like(param)
 
         self._op = (
             flow.builtin_op("adam_update")
